requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==6.3.0
phonenumbers==8.12.51
djangorestframework==3.13.1
requests==2.28.1
geopy==2.2.0
rollbar==0.16.3
GitPython==3.1.27